Refactored the reasoner
Previously the only supported reasoner was an HermiTReasoner, I have refactored
this to be any generic OWLReasoner.

The idea is that you create an OWLReasonerFactory that inherits from that
interface and then implement the instantiate reasoner methode that will take an
OWLOntology and return an OWLReasoner.

There is already a factory implementation in the code the HermiTReasonerFactory
that will instantiate the HermiTReasoner that was there before.

The ComplianceService instantiates the ComplianceChecker with an
OWLReasonerFactory at the moment this is hard coded to said
HermiTReasonerFactory, this will in a subsequent commit change to an
implementation with annotations and the pom.xml stating the configuration for
the OWLReasonerFactory.

Another point of note is that the explain function is no longer working, this is
mainly because in the previous implementation I depended on a functionality that
the HermiTReasoner offers. Another TODO is to figure out how to do that with
just the OWLAPI.

Action Items:
+ OWLReasonerFactory instantiation should be an annotation (automagic)
+ fix the explanation route with generic OWLAPI calls